+++ .eslintrc.json
@@ -9,8 +9,6 @@
   "rules": {
     "import/no-unresolved": "off",
     "no-restricted-globals": 0,
-    "no-var": 0,
-    "prefer-destructuring": 0,
     "prettier/prettier": [
       "error",
       {
@@ -22,5 +20,8 @@
   },
   "plugins": [
     "prettier"
-  ]
+  ],
+  "env": {
+    "browser": true
+  }
 }
