+++ package.json
@@ -4,10 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "echo 'just open the public folder!'",
-    "dev": "npx serve public",
-    "build:dev": "npm run build:css",
-    "build": "NODE_ENV=production npm run build:css",
+    "dev": "npm-run-all --parallel start watch",
+    "start": "serve public",
+    "build": "rollup -c",
+    "watch": "rollup -c -w",
+    "build-css:dev": "npm run build:css",
+    "build-css": "NODE_ENV=production npm run build:css",
     "build:css": "postcss src/stylesheets/style.css -o public/assets/style/style.css"
   },
   "keywords": [],
@@ -15,14 +17,24 @@
   "license": "ISC",
   "devDependencies": {
     "eslint-config-airbnb-base": "^14.0.0",
-    "postcss-cli": "^6.1.3"
+    "npm-run-all": "^4.1.5",
+    "postcss-cli": "^6.1.3",
+    "rollup": "^1.26.0",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
+    "rollup-plugin-postcss": "^2.0.3",
+    "rollup-plugin-terser": "^5.1.2",
+    "serve": "^11.2.0"
   },
   "dependencies": {
     "@fullhuman/postcss-purgecss": "^1.3.0",
     "autoprefixer": "^9.7.0",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
+    "classlist-polyfill": "^1.2.0",
     "cssnano": "^4.1.10",
+    "date-fns": "^2.6.0",
+    "es5-shim": "^4.5.13",
     "eslint": "^6.5.1",
     "eslint-config-prettier": "^6.4.0",
     "eslint-plugin-import": "^2.18.2",
