+++ public/index.html
@@ -33,12 +33,11 @@
   <meta name="theme-color" content="#ffffff">
   <link href="https://fonts.googleapis.com/css?family=Roboto:100,500,600&display=swap"
         rel="prefetch" media="all" as="style" onload="this.rel='stylesheet'">
-  <link href="assets/style/style.css" rel="stylesheet" media="all">
+  <link href="bundle.css" rel="stylesheet" media="all">
   <link rel="prefetch" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/3.7.2/animate.min.css" media="all"
         as="style" onload="this.rel='stylesheet'">
   <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=Intl.~locale.en" defer></script>
-  <script src="assets/scripts/polyfills/classList.polyfill.js" defer></script>
-  <script src="assets/scripts/main.js" defer></script>
+  <script src="bundle.js" defer></script>
 </head>
 <body class="flex h-full w-full bg-xmas">
 <div class="text-center lg:px-4 fixed w-full py-4 hidden animated shake" id="alert-no-date">
