+++ src/polyfills.js
@@ -1,0 +1,2 @@
+import 'classlist-polyfill';
+import 'es5-shim';
